Code cleanup: move and rename classes
